Optimized Circe decoder generation
